+++ lambdaface_face/src/Components/UserSettings.js
@@ -114,7 +114,7 @@ class UserSettings extends React.Component {
               id="email-input"
               className="user-settings__text-field"
               label="Email Address"
-          // className={}
+            // className={}
               type="email"
               value={this.state.email}
               onChange={this.handleChange("email")}
@@ -122,14 +122,14 @@ class UserSettings extends React.Component {
               required
             /> */}
             {/* <TextField
-          id="password-input"
-          label="Password"
-          // className={}
-          type="password"
-          value={this.state.password}
-          onChange={this.handleChange("password")}
-          margin="normal"
-        /> */}
+              id="password-input"
+              label="Password"
+              // className={}
+              type="password"
+              value={this.state.password}
+              onChange={this.handleChange("password")}
+              margin="normal"
+            /> */}
             <div className="user-settings__btns">
               <Button className="user-settings__reset-btn" variant="contained" color="primary" size="medium" onClick={this.resetPassword}>
               Reset Password
